+++ fractal_server/app/schemas/v2/__init__.py
@@ -31,6 +31,7 @@ from .task import TaskExportV2  # noqa F401
 from .task import TaskImportV2  # noqa F401
 from .task import TaskImportV2Legacy  # noqa F401
 from .task import TaskReadV2  # noqa F401
+from .task import TaskType  # noqa F401
 from .task import TaskUpdateV2  # noqa F401
 from .task_collection import TaskCollectCustomV2  # noqa F401
 from .task_collection import TaskCollectPipV2  # noqa F401