+++ fractal_server/app/schemas/v2/manifest.py
@@ -4,6 +4,7 @@ from pydantic import BaseModel
 from pydantic import Field
 from pydantic import model_validator
 
+from .task import TaskType
 from fractal_server.types import DictStrAny
 from fractal_server.types import HttpUrlStr
 from fractal_server.types import NonEmptyStr
@@ -55,15 +56,7 @@ class TaskManifestV2(BaseModel):
     modality: str | None = None
     tags: list[str] = Field(default_factory=list)
 
-    type: None | (
-        Literal[
-            "compound",
-            "converter_compound",
-            "non_parallel",
-            "converter_non_parallel",
-            "parallel",
-        ]
-    ) = None
+    type: None | TaskType = None
 
     @model_validator(mode="after")
     def validate_executable_args_meta(self):